refactor: Extract the debug option into a re-usable option
